requirements.txt
asgiref             ==3.6.0  
Django              ==4.1.4  
djangorestframework ==3.14.0 
pip                 ==22.3.1 
pytz                ==2022.7 
setuptools          ==65.5.0 
sqlparse            ==0.4.3  
tzdata              ==2022.7
python-decouple     ==3.6